Fix database connection to use utf8 as standard.
Currently the database connection uses "whatever" encoding for the connection. This means the encoding actually changed from server to server in my testing due to configuration differences.
So I fixed the encoding to be utf8 here. If there ever is the need for another encoding this must be made configurable.

Note: Magento also uses utf8 everywhere.